+++ src/control/trajectory_point_interface.cpp
@@ -277,8 +277,7 @@ void TrajectoryPointInterface::disconnectionCallback(const socket_t filedescript
   URCL_LOG_DEBUG("Connection to trajectory interface dropped.");
   for (auto handler : disconnect_callbacks_)
   {
-    // socket_t is UINT_PTR on Windows, so a narrowing cast is needed to match the callback signature
-    handler.function(static_cast<int>(filedescriptor));
+    handler.function(filedescriptor);
   }
   client_fd_ = INVALID_SOCKET;
 }
